fix: reconnection socketId changing issue resolved
